% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{download_latest_cash_file}
\alias{download_latest_cash_file}
\title{Descarga de telegram el ultimo archivo de pasta}
\usage{
download_latest_cash_file()
}
\value{

}
\description{
Descarga de telegram el ultimo archivo de pasta
}
