% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_book_accounts}
\alias{get_book_accounts}
\title{Get a list of the book accounts}
\usage{
get_book_accounts(...)
}
\arguments{
\item{xml_book}{}
}
\value{
tibble
}
\description{
Get a list of the book accounts
}
