% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_book_currencies}
\alias{get_book_currencies}
\title{Get a list of the books currencies}
\usage{
get_book_currencies(xml_book)
}
\arguments{
\item{xml_book}{}
}
\value{
tibble
}
\description{
Get a list of the books currencies
}
