% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_book_metadata}
\alias{get_book_metadata}
\title{Get an specific book metadata}
\usage{
get_book_metadata(xml_book)
}
\arguments{
\item{xml_book}{}
}
\value{

}
\description{
Get an specific book metadata
}
\examples{
get_book_metadata(book)
}
