% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_book_transactions}
\alias{get_book_transactions}
\title{Get all of the account transactions}
\usage{
get_book_transactions(...)
}
\arguments{
\item{xml_book}{}
}
\value{
tibble
}
\description{
Get all of the account transactions
}
