% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_books}
\alias{get_books}
\title{Return books from file}
\usage{
get_books(xml_root_node)
}
\arguments{
\item{xml_root_node}{}
}
\value{
tibble
}
\description{
Return books from file
}
\examples{
get_books(xml)
}
