% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{get_parents_list}
\alias{get_parents_list}
\title{Get parents and add their names}
\usage{
get_parents_list(accounts)
}
\arguments{
\item{accounts}{}
}
\value{

}
\description{
Get parents and add their names
}
