% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_date.R
\name{mydateInput}
\alias{mydateInput}
\title{Custom input date}
\usage{
mydateInput(
  inputId,
  label,
  value = NULL,
  min = NULL,
  max = NULL,
  format = "yyyy-mm-dd",
  startview = "month",
  weekstart = 0,
  minviewmode = "months",
  language = "en",
  width = NULL,
  autoclose = TRUE,
  datesdisabled = NULL,
  daysofweekdisabled = NULL
)
}
\arguments{
\item{daysofweekdisabled}{}
}
\value{

}
\description{
Custom input date
}
