% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cash_csv.R
\name{read_cash_csv}
\alias{read_cash_csv}
\title{Lectura de datos (csv)}
\usage{
read_cash_csv(file_list)
}
\arguments{
\item{file_list}{}
}
\value{
tibble
}
\description{
Lectura de datos (csv)
}
\examples{
read_cash_csv(file_list)
}
