% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{read_cash_xml}
\alias{read_cash_xml}
\title{Read xml cash file}
\usage{
read_cash_xml(file)
}
\arguments{
\item{file}{}
}
\value{
tibble
}
\description{
Read xml cash file
}
\examples{
read_cash_xml(file)
}
